+++ src/components/AbstractSelectInput/AbstractSelectInput.js
@@ -1,5 +1,6 @@
-import PropTypes from 'prop-types'
 import { Component } from 'react'
+
+import propTypes from './types.js'
 
 class AbstractSelectInput extends Component {
   constructor(props) {
@@ -56,13 +57,6 @@ class AbstractSelectInput extends Component {
   }
 }
 
-AbstractSelectInput.propTypes = {
-  value: PropTypes.any,
-  onEndEditing: PropTypes.func,
-  onSubmitEditing: PropTypes.func,
-  onValueChange: PropTypes.func,
-  options: PropTypes.array,
-  value: PropTypes.any
-}
+AbstractSelectInput.propTypes = propTypes
 
 export default AbstractSelectInput
