+++ src/components/AbstractSelectInput/index.js
@@ -1,0 +1,1 @@
+export { default } from './AbstractSelectInput.js'
