+++ src/components/AbstractSelectInput/types.js
@@ -1,0 +1,10 @@
+import PropTypes from 'prop-types'
+
+export default {
+  value: PropTypes.any,
+  onEndEditing: PropTypes.func,
+  onSubmitEditing: PropTypes.func,
+  onValueChange: PropTypes.func,
+  options: PropTypes.array,
+  value: PropTypes.any
+}
