+++ src/components/CustomKeyboard/styles.js
