+++ src/components/CustomKeyboard/types.js
@@ -1,0 +1,15 @@
+import PropTypes from 'prop-types'
+
+export default {
+  buttonsBackgroundColor: PropTypes.string,
+  buttonsBorderColor: PropTypes.string,
+  buttonsBorderWidth: PropTypes.number,
+  buttonsTextColor: PropTypes.string,
+  buttonsTextSize: PropTypes.number,
+  cancelKeyText: PropTypes.string,
+  onCancelPress: PropTypes.func.isRequired,
+  onSubmitPress: PropTypes.func.isRequired,
+  submitKeyText: PropTypes.string,
+  visible: PropTypes.bool.isRequired,
+  children: PropTypes.element
+}
