+++ src/components/KeyboardButton/index.js
@@ -1,0 +1,1 @@
+export { default } from './KeyboardButton.js'
