+++ src/components/KeyboardButton/types.js
@@ -1,0 +1,9 @@
+import PropTypes from 'prop-types'
+
+export default {
+  color: PropTypes.string.isRequired,
+  onPress: PropTypes.func.isRequired,
+  text: PropTypes.string.isRequired,
+  textAlign: PropTypes.string.isRequired,
+  textSize: PropTypes.number
+}
