+++ src/components/PickerItems/types.js
@@ -1,0 +1,5 @@
+import PropTypes from 'prop-types'
+
+export default {
+  options: PropTypes.array
+}
