+++ src/components/SelectInput.android/index.js
@@ -1,0 +1,1 @@
+export { default } from './SelectInput.js'
