+++ src/components/SelectInput.android/styles.js
