+++ src/components/SelectInput.ios/index.js
@@ -1,0 +1,1 @@
+export { default } from './SelectInput.js'
