+++ src/components/SelectInput.ios/styles.js
